Made some refactoring as per Github comments.
